+++ src/routes/index.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const userController = require('../controllers/UserController');
 const authController = require('../controllers/AuthController.js');
+const messageController = require('../controllers/MessageController');
 const passport = require('passport');
 const userRegistValidator = require('../validators/userRegistValidator');
 const messageRegistValidator = require('../validators/messageRegistValidoatr');
@@ -20,9 +21,9 @@ router.get('/user', userController.myPage);
 router.get('/register', userController.register);
 router.post('/signup', userRegistValidator, authController.postUser);
 router.post('/logout', authController.logout);
-router.get('/message', userController.message);
-router.get('/messages/:id', userController.message);
-router.post('/messages/:id/update', userController.updateMessage)
-router.post('/message', messageRegistValidator, userController.postMessage);
-router.get('/messages/:id/delete', userController.deleteMessage)
+router.get('/message', messageController.message);
+router.get('/messages/:id', messageController.message);
+router.post('/messages/:id/update', messageController.updateMessage)
+router.post('/message', messageRegistValidator, messageController.postMessage);
+router.get('/messages/:id/delete', messageController.deleteMessage)
 module.exports = router;
